+++ src/model/data/mod.rs
@@ -2,6 +2,8 @@ pub mod field;
 pub mod maker;
 pub mod map;
 pub mod record;
+pub mod storage;
 pub use field::Field;
 pub use map::ObjectMap;
 pub use record::Record;
+pub use storage::FieldStorage;